maint(core): Bring more validation into the parse config function
