+++ packages/plugins/src/hardhat/deployments.ts
@@ -10,14 +10,14 @@ import {
   writeSnapshotId,
   resolveNetworkName,
   ChugSplashExecutorType,
-  readUserChugSplashConfig,
   getDefaultProxyAddress,
-  readParsedChugSplashConfig,
+  readUnvalidatedChugSplashConfig,
 } from '@chugsplash/core'
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 
-import { getArtifactPaths, importOpenZeppelinStorageLayouts } from './artifacts'
+import { getArtifactPaths } from './artifacts'
 import { isRemoteExecution } from './utils'
+import { createChugSplashRuntime } from '../utils'
 
 export const fetchFilesRecursively = (dir): string[] => {
   const paths: string[] = []
@@ -45,8 +45,6 @@ export const deployAllChugSplashConfigs = async (
   hre: HardhatRuntimeEnvironment,
   silent: boolean,
   ipfsUrl: string,
-  noCompile: boolean,
-  confirm: boolean,
   fileNames?: string[]
 ) => {
   const remoteExecution = await isRemoteExecution(hre)
@@ -62,30 +60,24 @@ export const deployAllChugSplashConfigs = async (
   const deploymentFolder = hre.config.paths.deployments
 
   for (const configPath of fileNames) {
+    const cre = await createChugSplashRuntime(
+      configPath,
+      remoteExecution,
+      true,
+      hre
+    )
+
     // Skip this config if it's empty.
     if (isEmptyChugSplashConfig(configPath)) {
       return
     }
-    const userConfig = await readUserChugSplashConfig(configPath)
+    const userConfig = await readUnvalidatedChugSplashConfig(configPath)
 
     const artifactPaths = await getArtifactPaths(
       hre,
       userConfig.contracts,
       hre.config.paths.artifacts,
       path.join(hre.config.paths.artifacts, 'build-info')
-    )
-
-    const parsedConfig = await readParsedChugSplashConfig(
-      hre.ethers.provider,
-      configPath,
-      artifactPaths,
-      'hardhat'
-    )
-
-    const openzeppelinStorageLayouts = await importOpenZeppelinStorageLayouts(
-      hre,
-      parsedConfig,
-      userConfig
     )
 
     const signer = hre.ethers.provider.getSigner()
@@ -96,8 +88,6 @@ export const deployAllChugSplashConfigs = async (
       silent,
       remoteExecution,
       ipfsUrl,
-      noCompile,
-      confirm,
       true,
       await signer.getAddress(),
       false,
@@ -105,9 +95,8 @@ export const deployAllChugSplashConfigs = async (
       canonicalConfigPath,
       deploymentFolder,
       'hardhat',
-      true,
-      executor,
-      openzeppelinStorageLayouts
+      cre,
+      executor
     )
   }
 }
@@ -128,7 +117,7 @@ export const getContract = async (
 
   const resolvedConfigs = await Promise.all(
     filteredConfigNames.map((configFileName) => {
-      return readUserChugSplashConfig(configFileName)
+      return readUnvalidatedChugSplashConfig(configFileName)
     })
   )
