+++ packages/plugins/src/hardhat/tasks.ts
@@ -16,7 +16,6 @@ import {
   initializeChugSplash,
   monitorChugSplashSetup,
   chugsplashRegisterAbstractTask,
-  readParsedChugSplashConfig,
   chugsplashCommitAbstractSubtask,
   bundleLocal,
   chugsplashProposeAbstractTask,
@@ -36,8 +35,9 @@ import {
   ChugSplashExecutorType,
   ArtifactPaths,
   bundleRemoteSubtask,
-  readUserChugSplashConfig,
   ChugSplashBundles,
+  readValidatedChugSplashConfig,
+  readUnvalidatedChugSplashConfig,
 } from '@chugsplash/core'
 import { ChugSplashManagerABI, EXECUTOR } from '@chugsplash/contracts'
 import ora from 'ora'
@@ -54,8 +54,9 @@ import {
   sampleTestFileJavaScript,
   sampleTestFileTypeScript,
 } from '../sample-project/sample-tests'
-import { getArtifactPaths, importOpenZeppelinStorageLayouts } from './artifacts'
+import { getArtifactPaths } from './artifacts'
 import { isRemoteExecution } from './utils'
+import { createChugSplashRuntime } from '../utils'
 
 // Load environment variables from .env
 dotenv.config()
@@ -127,7 +128,6 @@ export const chugsplashDeployTask = async (
     noCompile: boolean
     confirm: boolean
     noWithdraw: boolean
-    skipStorageCheck: boolean
   },
   hre: HardhatRuntimeEnvironment
 ) => {
@@ -140,8 +140,14 @@ export const chugsplashDeployTask = async (
     noCompile,
     confirm,
     noWithdraw,
-    skipStorageCheck,
   } = args
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    confirm,
+    hre
+  )
 
   if (!noCompile) {
     await hre.run(TASK_COMPILE, {
@@ -155,8 +161,6 @@ export const chugsplashDeployTask = async (
   const signer = hre.ethers.provider.getSigner()
   const signerAddress = await signer.getAddress()
   await initializeChugSplash(hre.ethers.provider, signer, [signerAddress])
-
-  const remoteExecution = await isRemoteExecution(hre)
 
   let executor: ChugSplashExecutorType | undefined
   if (remoteExecution) {
@@ -172,25 +176,12 @@ export const chugsplashDeployTask = async (
   const canonicalConfigPath = hre.config.paths.canonicalConfigs
   const deploymentFolder = hre.config.paths.deployments
 
-  const userConfig = await readUserChugSplashConfig(configPath)
+  const userConfig = await readUnvalidatedChugSplashConfig(configPath)
   const artifactPaths = await getArtifactPaths(
     hre,
     userConfig.contracts,
     hre.config.paths.artifacts,
     path.join(hre.config.paths.artifacts, 'build-info')
-  )
-
-  const parsedConfig = await readParsedChugSplashConfig(
-    hre.ethers.provider,
-    configPath,
-    artifactPaths,
-    'hardhat'
-  )
-
-  const openzeppelinStorageLayouts = await importOpenZeppelinStorageLayouts(
-    hre,
-    parsedConfig,
-    userConfig
   )
 
   await chugsplashDeployAbstractTask(
@@ -200,8 +191,6 @@ export const chugsplashDeployTask = async (
     silent,
     remoteExecution,
     ipfsUrl,
-    noCompile,
-    confirm,
     !noWithdraw,
     newOwner ?? signerAddress,
     allowManagedProposals,
@@ -209,9 +198,8 @@ export const chugsplashDeployTask = async (
     canonicalConfigPath,
     deploymentFolder,
     'hardhat',
-    skipStorageCheck,
-    executor,
-    openzeppelinStorageLayouts
+    cre,
+    executor
   )
 }
 
@@ -240,10 +228,6 @@ task(TASK_CHUGSPLASH_DEPLOY)
     'allowManagedProposals',
     'Allow the ChugSplash Managed Service to propose deployments and upgrades on your behalf.'
   )
-  .addFlag(
-    'skipStorageCheck',
-    "Upgrade your contract(s) without checking for storage layout compatibility. Only use this when confident that the upgrade won't lead to storage layout issues."
-  )
   .setAction(chugsplashDeployTask)
 
 export const chugsplashRegisterTask = async (
@@ -256,13 +240,20 @@ export const chugsplashRegisterTask = async (
   hre: HardhatRuntimeEnvironment
 ) => {
   const { configPath, silent, owner, allowManagedProposals } = args
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    true,
+    hre
+  )
 
   const provider = hre.ethers.provider
   const signer = hre.ethers.provider.getSigner()
   const signerAddress = await signer.getAddress()
   await initializeChugSplash(hre.ethers.provider, signer, [signerAddress])
 
-  const userConfig = await readUserChugSplashConfig(configPath)
+  const userConfig = await readUnvalidatedChugSplashConfig(configPath)
   const artifactPaths = await getArtifactPaths(
     hre,
     userConfig.contracts,
@@ -270,11 +261,12 @@ export const chugsplashRegisterTask = async (
     path.join(hre.config.paths.artifacts, 'build-info')
   )
 
-  const parsedConfig = await readParsedChugSplashConfig(
+  const parsedConfig = await readValidatedChugSplashConfig(
     provider,
     configPath,
     artifactPaths,
-    'hardhat'
+    'hardhat',
+    cre
   )
 
   await chugsplashRegisterAbstractTask(
@@ -306,12 +298,17 @@ export const chugsplashProposeTask = async (
     silent: boolean
     noCompile: boolean
     confirm: boolean
-    skipStorageCheck: boolean
   },
   hre: HardhatRuntimeEnvironment
 ) => {
-  const { configPath, ipfsUrl, silent, noCompile, confirm, skipStorageCheck } =
-    args
+  const { configPath, ipfsUrl, silent, noCompile, confirm } = args
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    confirm,
+    hre
+  )
 
   if (!noCompile) {
     await hre.run(TASK_COMPILE, {
@@ -319,13 +316,13 @@ export const chugsplashProposeTask = async (
     })
   }
 
+  const userConfig = await readUnvalidatedChugSplashConfig(configPath)
+  const canonicalConfigPath = hre.config.paths.canonicalConfigs
+
   const provider = hre.ethers.provider
   const signer = hre.ethers.provider.getSigner()
   const signerAddress = await signer.getAddress()
   await initializeChugSplash(hre.ethers.provider, signer, [signerAddress])
-
-  const userConfig = await readUserChugSplashConfig(configPath)
-  const canonicalConfigPath = hre.config.paths.canonicalConfigs
 
   const artifactPaths = await getArtifactPaths(
     hre,
@@ -334,36 +331,25 @@ export const chugsplashProposeTask = async (
     path.join(hre.config.paths.artifacts, 'build-info')
   )
 
-  const parsedConfig = await readParsedChugSplashConfig(
+  const parsedConfig = await readValidatedChugSplashConfig(
     provider,
     configPath,
     artifactPaths,
-    'hardhat'
-  )
-
-  const remoteExecution = await isRemoteExecution(hre)
-
-  const openzeppelinStorageLayouts = await importOpenZeppelinStorageLayouts(
-    hre,
-    parsedConfig,
-    userConfig
+    'hardhat',
+    cre
   )
 
   await chugsplashProposeAbstractTask(
     provider,
     signer,
     parsedConfig,
-    userConfig,
     configPath,
     ipfsUrl,
     silent,
     remoteExecution,
-    confirm,
     'hardhat',
     artifactPaths,
-    canonicalConfigPath,
-    skipStorageCheck,
-    openzeppelinStorageLayouts
+    canonicalConfigPath
   )
 }
 
@@ -380,10 +366,6 @@ task(TASK_CHUGSPLASH_PROPOSE)
     'confirm',
     'Automatically confirm contract upgrades. Only applicable if upgrading on a live network.'
   )
-  .addFlag(
-    'skipStorageCheck',
-    "Upgrade your contract(s) without checking for storage layout compatibility. Only use this when confident that the upgrade won't lead to storage layout issues."
-  )
   .setAction(chugsplashProposeTask)
 
 export const chugsplashApproveTask = async (
@@ -397,17 +379,22 @@ export const chugsplashApproveTask = async (
 ) => {
   const { configPath, noWithdraw, silent, skipMonitorStatus } = args
 
+  const canonicalConfigPath = hre.config.paths.canonicalConfigs
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    true,
+    hre
+  )
+
+  const userConfig = await readUnvalidatedChugSplashConfig(configPath)
   const provider = hre.ethers.provider
   const signer = hre.ethers.provider.getSigner()
   const signerAddress = await signer.getAddress()
   await initializeChugSplash(hre.ethers.provider, signer, [signerAddress])
 
-  const canonicalConfigPath = hre.config.paths.canonicalConfigs
   const deploymentFolder = hre.config.paths.deployments
-
-  const remoteExecution = await isRemoteExecution(hre)
-
-  const userConfig = await readUserChugSplashConfig(configPath)
   const artifactPaths = await getArtifactPaths(
     hre,
     userConfig.contracts,
@@ -426,7 +413,8 @@ export const chugsplashApproveTask = async (
     'hardhat',
     canonicalConfigPath,
     deploymentFolder,
-    remoteExecution
+    remoteExecution,
+    cre
   )
 }
 
@@ -502,11 +490,9 @@ export const chugsplashCommitSubtask = async (
   }
 
   const canonicalConfigPath = hre.config.paths.canonicalConfigs
-
   const provider = hre.ethers.provider
   return chugsplashCommitAbstractSubtask(
     provider,
-    provider.getSigner(),
     parsedConfig,
     ipfsUrl,
     commitToIpfs,
@@ -655,18 +641,22 @@ export const monitorTask = async (
   hre: HardhatRuntimeEnvironment
 ) => {
   const { configPath, noWithdraw, silent, newOwner } = args
+  const canonicalConfigPath = hre.config.paths.canonicalConfigs
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    true,
+    hre
+  )
 
   const provider = hre.ethers.provider
   const signer = hre.ethers.provider.getSigner()
   const signerAddress = await signer.getAddress()
   await initializeChugSplash(hre.ethers.provider, signer, [signerAddress])
 
-  const canonicalConfigPath = hre.config.paths.canonicalConfigs
   const deploymentFolder = hre.config.paths.deployments
-
-  const remoteExecution = await isRemoteExecution(hre)
-
-  const userConfig = await readUserChugSplashConfig(configPath)
+  const userConfig = await readUnvalidatedChugSplashConfig(configPath)
   const artifactPaths = await getArtifactPaths(
     hre,
     userConfig.contracts,
@@ -685,7 +675,8 @@ export const monitorTask = async (
     canonicalConfigPath,
     deploymentFolder,
     'hardhat',
-    remoteExecution
+    remoteExecution,
+    cre
   )
 }
 
@@ -708,13 +699,20 @@ export const chugsplashFundTask = async (
   hre: HardhatRuntimeEnvironment
 ) => {
   const { amount, silent, configPath, autoEstimate } = args
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    true,
+    hre
+  )
 
   const provider = hre.ethers.provider
   const signer = hre.ethers.provider.getSigner()
   const signerAddress = await signer.getAddress()
   await initializeChugSplash(hre.ethers.provider, signer, [signerAddress])
 
-  const userConfig = await readUserChugSplashConfig(configPath)
+  const userConfig = await readUnvalidatedChugSplashConfig(configPath)
   const artifactPaths = await getArtifactPaths(
     hre,
     userConfig.contracts,
@@ -730,7 +728,8 @@ export const chugsplashFundTask = async (
     autoEstimate,
     silent,
     artifactPaths,
-    'hardhat'
+    'hardhat',
+    cre
   )
 }
 
@@ -783,7 +782,7 @@ task(TASK_NODE)
               quiet: true,
             })
           }
-          await deployAllChugSplashConfigs(hre, hide, '', true, true)
+          await deployAllChugSplashConfigs(hre, hide, '')
           const networkName = await resolveNetworkName(
             hre.ethers.provider,
             'hardhat'
@@ -822,8 +821,9 @@ task(TASK_TEST)
       runSuper
     ) => {
       const { show, noCompile, configPath, skipDeploy } = args
-      const signer = hre.ethers.provider.getSigner()
       const remoteExecution = await isRemoteExecution(hre)
+
+      const signer = hre.ethers.provider.getSigner()
       const executor = remoteExecution ? EXECUTOR : await signer.getAddress()
       const networkName = await resolveNetworkName(
         hre.ethers.provider,
@@ -856,8 +856,6 @@ task(TASK_TEST)
               hre,
               !show,
               '',
-              true,
-              true,
               configPath ? [configPath] : undefined
             )
           }
@@ -892,11 +890,11 @@ task(TASK_RUN)
       runSuper
     ) => {
       const { deployAll, noCompile } = args
+      const remoteExecution = await isRemoteExecution(hre)
+
       if (deployAll) {
         const signer = hre.ethers.provider.getSigner()
 
-        const remoteExecution = await isRemoteExecution(hre)
-        const confirm = remoteExecution ? args.confirm : true
         const executor = remoteExecution ? EXECUTOR : await signer.getAddress()
         await initializeChugSplash(hre.ethers.provider, signer, [executor])
         if (!noCompile) {
@@ -904,7 +902,7 @@ task(TASK_RUN)
             quiet: true,
           })
         }
-        await deployAllChugSplashConfigs(hre, true, '', true, confirm)
+        await deployAllChugSplashConfigs(hre, true, '')
       }
       await runSuper(args)
     }
@@ -921,21 +919,7 @@ export const chugsplashCancelTask = async (
   const provider = hre.ethers.provider
   const signer = provider.getSigner()
 
-  const config = await readUserChugSplashConfig(configPath)
-  const artifactPaths = await getArtifactPaths(
-    hre,
-    config.contracts,
-    hre.config.paths.artifacts,
-    path.join(hre.config.paths.artifacts, 'build-info')
-  )
-
-  await chugsplashCancelAbstractTask(
-    provider,
-    signer,
-    configPath,
-    artifactPaths,
-    'hardhat'
-  )
+  await chugsplashCancelAbstractTask(provider, signer, configPath, 'hardhat')
 }
 
 task(TASK_CHUGSPLASH_CANCEL)
@@ -954,23 +938,12 @@ export const chugsplashWithdrawTask = async (
 
   const provider = hre.ethers.provider
   const signer = provider.getSigner()
-  const canonicalConfigPath = hre.config.paths.canonicalConfigs
-
-  const userConfig = await readUserChugSplashConfig(configPath)
-  const artifactPaths = await getArtifactPaths(
-    hre,
-    userConfig.contracts,
-    hre.config.paths.artifacts,
-    path.join(hre.config.paths.artifacts, 'build-info')
-  )
 
   await chugsplashWithdrawAbstractTask(
     provider,
     signer,
     configPath,
     silent,
-    artifactPaths,
-    canonicalConfigPath,
     'hardhat'
   )
 }
@@ -1003,19 +976,10 @@ export const listProposersTask = async (
   const provider = hre.ethers.provider
   const signer = provider.getSigner()
 
-  const config = await readUserChugSplashConfig(configPath)
-  const artifactPaths = await getArtifactPaths(
-    hre,
-    config.contracts,
-    hre.config.paths.artifacts,
-    path.join(hre.config.paths.artifacts, 'build-info')
-  )
-
   await chugsplashListProposersAbstractTask(
     provider,
     signer,
     configPath,
-    artifactPaths,
     'hardhat'
   )
 }
@@ -1041,20 +1005,11 @@ export const addProposerTask = async (
   const provider = hre.ethers.provider
   const signer = provider.getSigner()
 
-  const config = await readUserChugSplashConfig(configPath)
-  const artifactPaths = await getArtifactPaths(
-    hre,
-    config.contracts,
-    hre.config.paths.artifacts,
-    path.join(hre.config.paths.artifacts, 'build-info')
-  )
-
   await chugsplashAddProposersAbstractTask(
     provider,
     signer,
     configPath,
     newProposers,
-    artifactPaths,
     'hardhat'
   )
 }
@@ -1078,10 +1033,18 @@ export const claimProxyTask = async (
   hre: HardhatRuntimeEnvironment
 ) => {
   const { configPath, referenceName, silent } = args
+  const remoteExecution = await isRemoteExecution(hre)
+  const cre = await createChugSplashRuntime(
+    configPath,
+    remoteExecution,
+    true,
+    hre
+  )
+
   const provider = hre.ethers.provider
   const signer = provider.getSigner()
 
-  const config = await readUserChugSplashConfig(configPath)
+  const config = await readUnvalidatedChugSplashConfig(configPath)
   const artifactPaths = await getArtifactPaths(
     hre,
     config.contracts,
@@ -1096,7 +1059,8 @@ export const claimProxyTask = async (
     referenceName,
     silent,
     artifactPaths,
-    'hardhat'
+    'hardhat',
+    cre
   )
 }
 
@@ -1124,16 +1088,9 @@ export const transferOwnershipTask = async (
   hre: HardhatRuntimeEnvironment
 ) => {
   const { configPath, proxy, silent } = args
+
   const provider = hre.ethers.provider
   const signer = provider.getSigner()
-
-  const config = await readUserChugSplashConfig(configPath)
-  const artifactPaths = await getArtifactPaths(
-    hre,
-    config.contracts,
-    hre.config.paths.artifacts,
-    path.join(hre.config.paths.artifacts, 'build-info')
-  )
 
   await chugsplashTransferOwnershipAbstractTask(
     provider,
@@ -1141,7 +1098,6 @@ export const transferOwnershipTask = async (
     configPath,
     proxy,
     silent,
-    artifactPaths,
     'hardhat'
   )
 }
@@ -1166,7 +1122,6 @@ export const chugsplashInitTask = async (
   hre: HardhatRuntimeEnvironment
 ) => {
   const { silent } = args
-
   const spinner = ora({ isSilent: silent })
   spinner.start('Initializing ChugSplash project...')
